Revert "virtio/vsock: fix leaks due to missing skb owner"
This reverts commit a6650e78c428841da5d68904475c5d177f3d11ee which is
commit f9d2b1e146e0f82f3d04629afd92698522058361 upstream.

It needs to be reverted, as the function virtio_transport_free_pkt() is
in the stable Android ABI, and this function is removed in the 6.1.63
kernel as part of a series of fixes and optimizations of the virtio
networking connection.  To preserve the ABI (and thereby keeping the
slower interface), revert the commit for now.

If this needs to come back in the future, if performance for this
interface needs to be resolved, it can be done so in an ABI-safe way.

Bug: 161946584
Change-Id: Ia3ece70e6ced0b779c911cb47aa133c2d378aa82
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>