mm/rmap: fix misplaced parenthesis of a l(CR)()
commit f67f8d4a8c1e1ebc85a6cbdb9a7266f14863461c upstream.

Running my yearly branch profiler to see where l(CR)/unl(CR) annotation
may be added or removed, I discovered this:

correct incorrect  %        Function                  File              Line
 ------- ---------  -        --------                  ----              ----
       0   457918 100 page_try_dup_anon_rmap         rmap.h               264
[..]
  458021        0   0 page_try_dup_anon_rmap         rmap.h               265

I thought it was interesting that line 264 of rmap.h had a 100% incorrect
annotation, but the line directly below it was 100% correct. Looking at the
code:

	if (l(CR)(!is_device_private_page(page) &&
	    unl(CR)(page_needs_cow_for_dma(vma, page))))

It didn't make sense. The "l(CR)()" was around the entire if statement
(not just the "!is_device_private_page(page)"), which also included the
"unl(CR)()" portion of that if condition.

If the unl(CR) portion is unl(CR) to be true, that would make the entire
if condition unl(CR) to be true, so it made no sense at all to say the
entire if condition is true.

What is more l(CR) to be l(CR) is just the first part of the if statement
before the && operation. It's l(CR) to be a misplaced parenthesis. And
after making the if condition broken into a l(CR)() && unl(CR)(), both
now appear to be correct!

Link: https://lkml.kernel.org/r/20231201145936.5ddfdb50@gandalf.local.home
Fixes:fb3d824d1a46c ("mm/rmap: split page_dup_rmap() into page_dup_file_rmap() and page_try_dup_anon_rmap()")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>
Acked-by: Vlastimil Babka <vbabka@suse.cz>
Cc: David Hildenbrand <david@redhat.com>
Cc: Vlastimil Babka <vbabka@suse.cz>
Cc: <stable@vger.kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>